h hh ll flags working. l – needs rework for wchar

+++ ft_printf/str_converter/ft_itoa_d.c
@@ -6,13 +6,13 @@
 /*   By: lbaela <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/25 12:53:01 by lbaela            #+#    #+#             */
-/*   Updated: 2021/06/14 19:10:56 by lbaela           ###   ########.fr       */
+/*   Updated: 2021/06/18 13:11:46 by lbaela           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	translate_nums(char *num, long int n, unsigned int len)
+void	translate_nums(char *num, long long n, unsigned int len)
 {
 	if (n < -9223372036854775807)
 	{
@@ -28,12 +28,12 @@ void	translate_nums(char *num, long int n, unsigned int len)
 	}
 }
 
-char	*ft_itoa_d(long int n, char	*sign)
+char	*ft_itoa_d(long long n, char	*sign)
 {
 	char				*num;
 	unsigned int		len;
-	unsigned long int	mod;
-	long int			sn;
+	unsigned long long	mod;
+	long long			sn;
 
 	mod = 10;
 	len = 1;
@@ -43,7 +43,7 @@ char	*ft_itoa_d(long int n, char	*sign)
 		*sign = '-';
 		sn = -1 * sn;
 	}
-	while (sn % mod != (unsigned long int) sn)
+	while (sn % mod != (unsigned long long) sn)
 	{
 		mod *= 10;
 		len++;
